refactor(ui): extract progress-overlay presentation from main.ts
Move showProgressOverlay/hideProgressOverlay/updateProgressDisplay into
src/ui/progress-overlay.ts. The overlay now takes the tile count as a
parameter instead of reading state.selectedTiles.size internally, so the
module has no application-state coupling. The download lifecycle handlers
(handleTileComplete/Complete/Error) stay in main.ts; they belong with the
download controller and will move in a later step.

Verified: typecheck clean, build OK, e2e 6 pass / 1 skip (download and
cancel tests directly exercise the overlay). Autoreview clean.
